+++ src/react-app-env.d.ts
@@ -1,1 +1,18 @@
 /// <reference types="react-scripts" />
+
+type RectSize = {
+  width: number;
+  height: number;
+};
+
+type BoxProps = {
+  key: number;
+  srcX: number;
+  srcY: number;
+  srcWidth: number;
+  srcHeight: number;
+  showX: number;
+  showY: number;
+  showWidth: number;
+  showHeight: number;
+};
